src/com/leetcode/leetcode_344_Reverse_String_Easy/Solution344.java
package com.leetcode.leetcode_344_Reverse_String_Easy;

public class Solution344 {
    public void reverseString_1(char[] s) {
        for (int i = 0; i < s.length / 2; i++) {
            char tmp = s[i];
            s[i] = s[s.length - i - 1];
            s[s.length - i - 1] = tmp;
        }
    }

    public void reverseString(char[] s) {
        int i = 0;
        int j = s.length - 1;
        while (i < j) {
            char tmp = s[i];
            s[i] = s[j];
            s[j] = tmp;
            i++;
            j--;
        }
    }
}
